refactor(auth): replace ``AuthorizationExtractor`` interface with`` BearerAuthorizationExtractor`` class - remove AuthorizationExtractor interface for simplicity - update BearerAuthorizationExtractor to standalone class - adjust CheckLoginInterceptor, AuthController, and LoginMemberArgumentResolver to use BearerAuthorizationExtractor
