+++ src/main/kotlin/ecommerce/auth/controller/AuthController.kt
@@ -1,8 +1,7 @@
 package ecommerce.auth.controller
 
-import ecommerce.auth.BearerAuthorizationExtractor
 import ecommerce.auth.exception.AuthorizationException
-import ecommerce.auth.extractor.AuthorizationExtractor
+import ecommerce.auth.extractor.BearerAuthorizationExtractor
 import ecommerce.auth.service.AuthService
 import ecommerce.member.dto.MemberResponse
 import ecommerce.member.dto.TokenRequest
@@ -24,7 +23,7 @@ import org.springframework.web.bind.annotation.RestController
 class AuthController(
     private val authService: AuthService,
 ) {
-    private val authorizationExtractor: AuthorizationExtractor<String> = BearerAuthorizationExtractor()
+    private val authorizationExtractor: BearerAuthorizationExtractor = BearerAuthorizationExtractor()
 
     @PostMapping("/register")
     fun register(
