--- src/main/kotlin/ecommerce/auth/extractor/AuthorizationExtractor.kt
@@ -1,11 +1,0 @@
-package ecommerce.auth.extractor
-
-import jakarta.servlet.http.HttpServletRequest
-
-interface AuthorizationExtractor<T> {
-    fun extract(request: HttpServletRequest): T
-
-    companion object {
-        const val AUTHORIZATION = "Authorization"
-    }
-}
