+++ src/main/kotlin/ecommerce/auth/extractor/BearerAuthorizationExtractor.kt
@@ -1,19 +1,22 @@
-package ecommerce.auth
+package ecommerce.auth.extractor
 
 import ecommerce.auth.exception.AuthorizationException
-import ecommerce.auth.extractor.AuthorizationExtractor
 import jakarta.servlet.http.HttpServletRequest
 import org.springframework.stereotype.Component
 
 @Component
-class BearerAuthorizationExtractor : AuthorizationExtractor<String> {
-    override fun extract(request: HttpServletRequest): String {
+class BearerAuthorizationExtractor {
+    fun extract(request: HttpServletRequest): String {
         val authHeader =
-            request.getHeader(AuthorizationExtractor.AUTHORIZATION)
+            request.getHeader(AUTHORIZATION)
                 ?: throw AuthorizationException("Missing Authorization header")
         if (!authHeader.startsWith("Bearer ")) {
             throw AuthorizationException("Authorization header must start with 'Bearer '")
         }
         return authHeader.substring(7)
+    }
+
+    companion object {
+        const val AUTHORIZATION = "Authorization"
     }
 }
