+++ src/main/kotlin/ecommerce/auth/resolver/LoginMemberArgumentResolver.kt
@@ -2,8 +2,10 @@ package ecommerce.auth.resolver
 
 import ecommerce.auth.annotation.LoginMember
 import ecommerce.auth.exception.AuthorizationException
+import ecommerce.auth.extractor.BearerAuthorizationExtractor
 import ecommerce.auth.service.AuthService
 import ecommerce.member.repository.MemberRepository
+import jakarta.servlet.http.HttpServletRequest
 import org.springframework.core.MethodParameter
 import org.springframework.stereotype.Component
 import org.springframework.web.bind.support.WebDataBinderFactory
@@ -15,6 +17,7 @@ import org.springframework.web.method.support.ModelAndViewContainer
 class LoginMemberArgumentResolver(
     private val authService: AuthService,
     private val memberRepository: MemberRepository,
+    private val authorizationExtractor: BearerAuthorizationExtractor,
 ) : HandlerMethodArgumentResolver {
     override fun supportsParameter(parameter: MethodParameter): Boolean {
         return parameter.hasParameterAnnotation(LoginMember::class.java)
@@ -26,9 +29,15 @@ class LoginMemberArgumentResolver(
         webRequest: NativeWebRequest,
         binderFactory: WebDataBinderFactory?,
     ): Any {
+        val request =
+            webRequest.getNativeRequest(HttpServletRequest::class.java)
+                ?: throw AuthorizationException("Invalid request")
         val token =
-            webRequest.getHeader("Authorization")?.removePrefix("Bearer ")
-                ?: throw AuthorizationException("Missing Authorization header")
+            try {
+                authorizationExtractor.extract(request)
+            } catch (e: AuthorizationException) {
+                throw AuthorizationException("Failed to extract token: ${e.message}")
+            }
         val memberResponse =
             try {
                 authService.findMemberByToken(token)
